update to use Spark 3.3 (#253)
* update to spark3.3

* upd ci

* fix scalastyle

* fix lint

* add shim layer


* format


* format & update ray dependency


* upd


* adjust class path


* fix


* fix placement group ci & upd pom


* fix


* cleanup
