+++ .github/workflows/raydp.yml
@@ -31,7 +31,7 @@ jobs:
       matrix:
         os: [ ubuntu-latest, macos-latest ]
         python-version: [3.7, 3.8, 3.9]
-        spark-version: [3.1.3, 3.2.1]
+        spark-version: [3.1.3, 3.2.1, 3.3.0]
 
     runs-on: ${{ matrix.os }}
