+++ core/raydp-main/src/main/scala/org/apache/spark/deploy/raydp/ApplicationDescription.scala
