+++ core/shims/pom.xml
@@ -1,0 +1,39 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+
+  <parent>
+    <groupId>com.intel</groupId>
+    <artifactId>raydp-parent</artifactId>
+    <version>0.6.0-SNAPSHOT</version>
+    <relativePath>../pom.xml</relativePath>
+  </parent>
+
+  <artifactId>raydp-shims</artifactId>
+  <name>RayDP Shims</name>
+  <packaging>pom</packaging>
+
+  <modules>
+    <module>common</module>
+    <module>spark321</module>
+    <module>spark330</module>
+  </modules>
+ 
+  <properties>
+    <scala.binary.version>2.12</scala.binary.version>
+    <scala.plugin.version>4.3.0</scala.plugin.version>
+    <scalatest-maven-plugin.version>3.2.2</scalatest-maven-plugin.version>
+  </properties>
+
+  <build>
+    <plugins>
+        <plugin>
+            <groupId>net.alchim31.maven</groupId>
+            <artifactId>scala-maven-plugin</artifactId>
+             <version>${scala.plugin.version}</version>
+        </plugin>
+    </plugins>
+  </build>
+</project>
