+++ python/raydp/spark/ray_cluster.py
@@ -74,13 +74,15 @@ class SparkCluster(Cluster):
         # Ray's jar. Without this, the driver produces INFO logs and the level cannot be changed.
         spark_home = os.environ.get("SPARK_HOME", os.path.dirname(pyspark.__file__))
         log4j_path = os.path.abspath(os.path.join(spark_home, "jars/slf4j-log4j*.jar"))
+        commons_path = os.path.abspath(os.path.join(spark_home, "jars/commons-lang3-*.jar"))
         try:
             extra_jars = [extra_conf["spark.jars"]]
         except KeyError:
             extra_jars = []
         extra_conf["spark.jars"] = ",".join(glob.glob(RAYDP_CP) + extra_jars)
         driver_cp_key = "spark.driver.extraClassPath"
-        driver_cp = ":".join(glob.glob(log4j_path) + glob.glob(RAYDP_CP) + glob.glob(RAY_CP))
+        driver_cp = ":".join(glob.glob(log4j_path) + glob.glob(commons_path)
+                + glob.glob(RAYDP_CP) + glob.glob(RAY_CP))
         if driver_cp_key in extra_conf:
             extra_conf[driver_cp_key] = driver_cp + ":" + extra_conf[driver_cp_key]
         else:
