+++ python/raydp/tests/test_spark_cluster.py
@@ -125,7 +125,7 @@ def test_ray_dataset_to_spark(spark_on_ray_small):
 
 def test_placement_group(ray_cluster):
     for pg_strategy in ["PACK", "STRICT_PACK", "SPREAD", "STRICT_SPREAD"]:
-        spark = raydp.init_spark("test_strategy", 1, 1, "500 M",
+        spark = raydp.init_spark(f"test_strategy_{pg_strategy}_1", 1, 1, "500 M",
                                  placement_group_strategy=pg_strategy)
         result = spark.range(0, 10, numPartitions=10).count()
         assert result == 10
@@ -137,7 +137,7 @@ def test_placement_group(ray_cluster):
         pg = ray.util.placement_group([{"CPU": 1, "memory": utils.parse_memory_size("500 M")}],
                                       strategy=pg_strategy)
         ray.get(pg.ready())
-        spark = raydp.init_spark("test_bundle", 1, 1, "500 M",
+        spark = raydp.init_spark(f"test_bundle_{pg_strategy}_2", 1, 1, "500 M",
                                  placement_group=pg,
                                  placement_group_bundle_indexes=[0])
         result = spark.range(0, 10, numPartitions=10).count()
@@ -147,7 +147,7 @@ def test_placement_group(ray_cluster):
         time.sleep(5)
 
         # w/ existing placement group w/o bundle indexes
-        spark = raydp.init_spark("test_bundle", 1, 1, "500 M",
+        spark = raydp.init_spark(f"test_bundle_{pg_strategy}_3", 1, 1, "500 M",
                                  placement_group=pg)
         result = spark.range(0, 10, numPartitions=10).count()
         assert result == 10
